requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.23

# NLP & ML
sentence-transformers==2.3.0
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.26.2

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# OCR (Optional)
easyocr==1.7.0
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# LLM Integration (Google Gemini)
google-generativeai==0.8.3
protobuf==5.29.3
